Address PR #163 review comments
1. Remove docs/color-editor-gui-plan.md from this branch (reviewer asked
   for it on its own branch — the plan doc is unrelated to the code PR).

2. Move event loop and window setup out of the example into a library
   function `run_color_editor()` in color_map_editor_ui.rs. The example
   main.rs is now a trivial three-line wrapper that calls through
   `common::color_editor_example_from_string`, matching the pattern used
   by all other examples in the repo.

3. Add TODO comment on paint_gradient_bar noting that the ColorMap is
   re-allocated every frame and should be cached once keyframe sets grow
   or refresh rates increase.

4. Make blit_preview_to_framebuffer robust to buffer/layout size
   mismatches: it now clips to min(buffer, pane) in both dimensions
   rather than assuming the buffer exactly matches PREVIEW_W x TOTAL_H.
   Also demoted from pub to private since run_color_editor encapsulates
   the call.
