+++ examples/common/mod.rs
@@ -5,7 +5,12 @@ use std::path::{Path, PathBuf};
 #[allow(dead_code)]
 use fractal_renderer::{
     cli::{color_swatch::generate_color_swatch, explore::explore_fractal, render::render_fractal},
-    core::file_io::FilePrefix,
+    core::{
+        color_map_editor_ui::{run_color_editor, PREVIEW_W, TOTAL_H},
+        file_io::FilePrefix,
+        image_utils::{create_buffer, Renderable},
+    },
+    fractals::{common::FractalParams, quadratic_map::QuadraticMap},
 };
 
 #[allow(dead_code)]
@@ -125,4 +130,30 @@ pub fn color_swatch_example_from_string(example_name: &str) {
             file_base: String::from("result"),
         },
     );
+}
+
+#[allow(dead_code)]
+pub fn color_editor_example_from_string(example_name: &str) {
+    let params_path = example_params_path(example_name);
+    let json_text = read_params_file_or_panic(example_name, &params_path);
+    let fractal_params: FractalParams =
+        parse_params_json_or_panic(example_name, &params_path, &json_text);
+
+    let (keyframes, preview_buffer) = match fractal_params {
+        FractalParams::Mandelbrot(params) => {
+            let kf = params.color_map.keyframes.clone();
+            let mut preview_params = (*params).clone();
+            preview_params.image_specification.resolution = [PREVIEW_W, TOTAL_H];
+            let renderer = QuadraticMap::new(preview_params);
+            let mut buf = create_buffer(image::Rgb([0u8, 0, 0]), &[PREVIEW_W, TOTAL_H]);
+            renderer.render_to_buffer(&mut buf);
+            (kf, buf)
+        }
+        _ => panic!(
+            "color_editor_example '{}' requires Mandelbrot params",
+            example_name
+        ),
+    };
+
+    run_color_editor(preview_buffer, keyframes).unwrap();
 }
